+++ kernel/core/datamanagement/dataaccess.cpp
@@ -437,7 +437,7 @@ void replaceDataEntities(string& sLine, const string& sEntity, MemoryManager& _d
 {
 	string sEntityOccurence = "";
 	string sEntityName = sEntity.substr(0, sEntity.length()-1);
-	unsigned int nPos = 0;
+	size_t nPos = 0;
 	bool bWriteStrings = false;
 	bool bWriteFileName = false;
 	bool isCluster = sEntity.back() == '{';
@@ -2454,7 +2454,7 @@ bool parser_CheckMultArgFunc(const string& sLeft, const string& sRight)
 	// examine the first non-whitespace character of the right side:
 	// - if it is a parenthesis, then simply break
 	// - if it is a comma, then try to find the next parenthesis. If it exists, it can be a compare-like function signature
-	for (unsigned int i = 0; i < sRight.length(); i++)
+	for (size_t i = 0; i < sRight.length(); i++)
 	{
 		if (sRight[i] != ' ')
 		{
