+++ kernel/core/datamanagement/memory.cpp
@@ -1070,7 +1070,7 @@ vector<string> Memory::getHeadLineElement(const VectorIndex& _vCol) const
 {
     vector<string> vHeadLines;
 
-    for (unsigned int i = 0; i < _vCol.size(); i++)
+    for (size_t i = 0; i < _vCol.size(); i++)
     {
         if (_vCol[i] < 0)
             continue;
@@ -1337,11 +1337,11 @@ void Memory::writeData(int _nLine, int _nCol, const std::string& sValue)
 ///
 /// \param _idx Indices&
 /// \param _dData mu::value_type*
-/// \param _nNum unsigned int
+/// \param _nNum size_t
 /// \return void
 ///
 /////////////////////////////////////////////////
-void Memory::writeData(Indices& _idx, mu::value_type* _dData, unsigned int _nNum)
+void Memory::writeData(Indices& _idx, mu::value_type* _dData, size_t _nNum)
 {
     int nDirection = LINES;
 
@@ -2392,7 +2392,7 @@ mu::value_type Memory::num(const VectorIndex& _vLine, const VectorIndex& _vCol) 
     _vLine.setOpenEndIndex(lines-1);
     _vCol.setOpenEndIndex(cols-1);
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2405,7 +2405,7 @@ mu::value_type Memory::num(const VectorIndex& _vLine, const VectorIndex& _vCol) 
             continue;
         }
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0 || _vLine[i] >= elems || !isValidElement(_vLine[i], _vCol[j]))
                 nInvalid++;
@@ -2438,7 +2438,7 @@ mu::value_type Memory::and_func(const VectorIndex& _vLine, const VectorIndex& _v
 
     double dRetVal = NAN;
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2448,7 +2448,7 @@ mu::value_type Memory::and_func(const VectorIndex& _vLine, const VectorIndex& _v
         if (!elems)
             continue;
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0)
                 continue;
@@ -2496,7 +2496,7 @@ mu::value_type Memory::or_func(const VectorIndex& _vLine, const VectorIndex& _vC
     _vLine.setOpenEndIndex(lines-1);
     _vCol.setOpenEndIndex(cols-1);
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2506,7 +2506,7 @@ mu::value_type Memory::or_func(const VectorIndex& _vLine, const VectorIndex& _vC
         if (!elems)
             continue;
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0)
                 continue;
@@ -2550,7 +2550,7 @@ mu::value_type Memory::xor_func(const VectorIndex& _vLine, const VectorIndex& _v
 
     bool isTrue = false;
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2560,7 +2560,7 @@ mu::value_type Memory::xor_func(const VectorIndex& _vLine, const VectorIndex& _v
         if (!elems)
             continue;
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0)
                 continue;
@@ -2620,7 +2620,7 @@ mu::value_type Memory::cnt(const VectorIndex& _vLine, const VectorIndex& _vCol) 
     else if (_vCol.size() > 1 && _vLine.size() == 1)
         vDimLen = size(_vLine, AppDir::LINES);
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2645,7 +2645,7 @@ mu::value_type Memory::cnt(const VectorIndex& _vLine, const VectorIndex& _vCol) 
             continue;
         }
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0 || _vLine[i] >= elems)
                 nInvalid++;
@@ -2865,7 +2865,7 @@ mu::value_type Memory::med(const VectorIndex& _vLine, const VectorIndex& _vCol) 
 
     vData.reserve(_vLine.size()*_vCol.size());
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2875,7 +2875,7 @@ mu::value_type Memory::med(const VectorIndex& _vLine, const VectorIndex& _vCol) 
         if (!elems)
             continue;
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0)
                 continue;
@@ -2935,7 +2935,7 @@ mu::value_type Memory::pct(const VectorIndex& _vLine, const VectorIndex& _vCol, 
     if (dPct.real() >= 1 || dPct.real() <= 0)
         return NAN;
 
-    for (unsigned int j = 0; j < _vCol.size(); j++)
+    for (size_t j = 0; j < _vCol.size(); j++)
     {
         if (_vCol[j] < 0)
             continue;
@@ -2945,7 +2945,7 @@ mu::value_type Memory::pct(const VectorIndex& _vLine, const VectorIndex& _vCol, 
         if (!elems)
             continue;
 
-        for (unsigned int i = 0; i < _vLine.size(); i++)
+        for (size_t i = 0; i < _vLine.size(); i++)
         {
             if (_vLine[i] < 0)
                 continue;
